[BNPL][Clank] Allow reshowing the TTF bottom sheet after dismissal
Currently, if the user dismisses the Touch To Fill bottom sheet and
reselects a credit card form field, then the keyboard appears. The
bottom sheet would not show again unless the user refreshes the page.

During a BNPL flow, the flow could fail because an issuer denies a loan,
or the user doesn't like the issuer options, etc. In cases like these,
the user should have the opportunity to reuse the bottom sheet.

This CL modifies the dismissal logic to allow the Touch To Fill bottom
sheet to be reshown if a user dismisses the bottom sheet during a BNPL
flow. A new parameter, `shouldReshow`, is added to the `OnDismissed`
method. This parameter is set to true when the user dismisses the bottom
sheet during a BNPL flow, which allows the bottom sheet to be reshown.

Demo: http://screencast/cast/NTkwOTkyNjk2NTY3Mzk4NHxjMTFkYjY4Mi0xZA
Design: go/autofill-clank-bnpl
Bug: 430575808
Change-Id: Ie3f81f4c436bfee53da867535adbd4d98cb39afa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7147205
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Fernando Ramirez <ferny@google.com>
Cr-Commit-Position: refs/heads/main@{#1550624}