[SyncTests] Enforce missing test requirement in transport mode too
In the recent https://crrev.com/c/7201860, a TODO was added to address a
temporary workaround because a specific test expectation was not passing
in transport mode and required investigation.

In this patch, this TODO is addressed by using additional waiting logic
via `WaitForNoPaymentsCustomerData()`. The need for this arises due to
the fact that `IsAutofillWalletImportEnabled()` reacts very quickly to
signout events because it internally uses
`SyncUserSettings::GetSelectedTypes()`. This leads to
`PaymentsDataManager::GetCreditCards()` exposing an empty list long
before `GetPaymentsCustomerData()` reacts to sign-out events.

Change-Id: Ice4099231f635a4887da18f0ddb772fd5c6873ae
Bug: 353425612
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7204924
Auto-Submit: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Mahmoud Rashad <mmrashad@google.com>
Reviewed-by: Mahmoud Rashad <mmrashad@google.com>
Cr-Commit-Position: refs/heads/main@{#1550663}