Add has_shown_ui_once flag to ReadAnythingController.
This is needed because the SidePanelCoordinator uses a
SidePanelEntryWaiter to determine when to show the SidePanel. The waiter
lets the SidePanelCoordinator know when RM is loaded and ready to be
displayed. This works when Reading Mode opens for the first time in the
Side Panel, because the WebUI calls ShowUI on its host
(https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/webui/top_chrome/webui_contents_wrapper.cc;l=265;drc=4497bff67a809a49fdb7f3fa7635329021000e2f).
For the RM SidePanel, SidePanelWebUIView is the host, so
SidePanelWebUIView::ShowUI() gets called when the WebUI loaded. This
tells the SidePanelContentProxy that the WebUI is "available", and this
gets propagated to the SidePanelCoordinator via the Waiter to show the
SidePanel.

However, when RM opens in the Side Panel after the WebUI already being
preloaded (e.g. when it's being transferred from Immersive Reading
Mode), it is loaded before the SidePanelWebUIView is the host, so this
whole process is skipped and the SidePanelCoordinator doesn't know that
the content is ready to get shown.

This change ensures that the ReadAnythingController tracks when the RM
WebUI has been shown, so the SidePanelWebUIView knows to call ShowUI if
the WebUI has been preloaded.

Bug: 460136293
Change-Id: If0497af0a0538e0c6dc61d6bd33018a64137a9ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7205125
Reviewed-by: Shari Trewin <trewin@google.com>
Commit-Queue: Eitan Goldberger <eitang@google.com>
Cr-Commit-Position: refs/heads/main@{#1550668}