[Contextual Tasks] Allow resumption based on task ID in URL
This patch updates the logic in GetUrlForTask to attempt to get the
URL for the AIM thread if getting the initial URL fails (now provided
as an optional). This results in both opening the task in a new tab
and restoring a session with a contextual tasks tab actually loading
content (pending support from the ContextualTasksService).

Bug: 460833892, 460657977
Change-Id: I8d07d97786fb570f09a0486c733525830e6a0f04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7153742
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1550669}