Fix PrefHashFilter crash on type mismatch in deferred revalidation
This CL fixes a crash in PrefHashFilter that occurs during deferred
revalidation when there is a type mismatch between the preference value
at load time and the registered preference type. The crash happens
because GetUserPrefValue() is called before the type check, leading to a
CHECK failure. The fix moves the GetUserPrefValue() call to after the
type check. A new test case is added to reproduce the crash and verify
the fix.

Bug: 462498564
Change-Id: Ic0b1daff0935e79b635806685bc6f5f5763b60b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7200263
Commit-Queue: Zack Han <zackhan@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Abe Boujane <boujane@google.com>
Cr-Commit-Position: refs/heads/main@{#1550689}