[ios] Fix proactive pwd generation sheet background hang
Do not refocus on the field if the sheet is dismissed when going in
the background.

Add actions so we can get more precise breadcrumbs of when the
hang happens.

Also fix the dropdown metric so the usage of the proactive bottom
sheet is also reported.

Bug: 40746470, 441363369
Change-Id: I2ef3a70fa0c7e24b21df4c599daf28626a8a9393
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7200364
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1550710}