[Prompt API WPTs] Update `inputUsage` to reflect total
tokens used in AI Language Model.

The `inputUsage` field in `AILanguageModelInstanceInfo` and
`ModelExecutionContextInfo` now reports the total number of tokens
consumed by the session, including both the initial prompts and all
subsequent inputs and outputs. This is achieved by tracking the initial
token count in the `Context` and summing it with the tokens from
subsequent interactions.

Bug: 460794950
Change-Id: I5b98e86d6795d4b1463b91b12d3f2e7c6751c8c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7170679
Reviewed-by: Nathan Memmott <memmott@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Isaac Ahouma <iahouma@google.com>
Cr-Commit-Position: refs/heads/main@{#1550722}