spanification: Clean up unneeded parameters in remoting
This was the result of asking gemini to remove all of the redundant
variables and update call sites. I have also gone ahead and updated all
the for loops that the linter complained were not modern c++ (to range
for loops).

The aim of this patch should be no behaviour changes, however well I've
manually reviewed it, since this patch was generated by AI please review
carefully.

These checks have been in the code for 2 stable releases, and M143 is
about to go to Stable, which means this has been fully fatal in
Canary/Dev. All of these should be completely redundant.

This patch was split with `git cl split --target-range 8 32
--max-depth=2 --no-reviewers -d ~/description.txt`

This is part of auto spanification see internal doc about it:
https://docs.google.com/document/u/1/d/1poEhNARfFas2osqPoXGzVZtBgnOQGQtQNUFTA-MlADs/edit?tab=t.0

This CL was uploaded by an experimental version of git cl split
(https://crbug.com/389069356).

Bug: 431824301
Change-Id: Iaf70332b8a36e310b776f98163931a5a7f7b5b78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7201765
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550731}