[Bookmarks Bar] Fix the scrolling animation of Bookmark Bar on the NTP
This CL addresses the issue of the Bookmark Bar not handling the yOffset
of widgets when the TopControlsStacker provides them, and BCIV is not
able to handle the scroll. This happens in cases such as on the NTP
when the width of a window is changed and the tab strip needs to be
hidden. We manually translate the Android widget on the y-axis based
on the provided yOffset, and ensure that the Android widget's visibility
is not updated (as is usually the case on non-zero translation for the
top control).

Bug: 454114987
Change-Id: Iad75c3a1909879ebba1599825d35de4e2479724b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7186755
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1550733}