[TBHSP] Update shadow opacity to correct value on close
Fixes a bug where the shadow opacity was always set to 100% when the
animation ended. This caused a bug during the side panel close where you
could see the shadow outline displayed in the tabstrip which could be
seen after the shadow animation has finished but before the side panel
close animation finishes.

To fix this, OnAnimationSequenceEnded reacts to the closing state of the
animation coordinator to verify the final shadow opacity.

Change-Id: I150db880e71ad7153e8320b0c533197956570936
Bug: 445452247
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7208288
Commit-Queue: Darryl James <dljames@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550794}