[crd host][wayland] PortalDesktopResizer & PortalDisplayInfoLoader
This is pretty much what we did in GnomeDesktopResizer and
GnomeDisplayInfoLoader prior to the integration with the GNOME
DisplayConfig API. They just simply use the PipeWire stream for resizing
and getting the resolution. Also, since there is no universal display
changed signal in Wayland, we will be polling the display info for
the time being.

Bug: 445973705
Change-Id: If020c94a5388eeffdbcb267fb08afeae6412647b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7208457
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550855}