Fix IsolatedWebAppRetryTest.RetryTimeStepsCorrect test
The test was failing because the
`IsolatedWebAppPolicyManager::DoProcessPolicy` was called multiple
times from the first run loop, usually with only the last invocation
having the IsolatedWebAppInstallForceList filled out, but not always.

The unwanted invocations were leftovers from setup. Now we clear them
out with a run loop, so when the IsolatedWebAppInstallForceList is
updated, the DoProcessPolicy is called just once.

Bug: 445714668
Change-Id: I9fada00339c1e2d7590c15eaafa95074633a8d79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7207118
Commit-Queue: Patryk Chodur <pchodur@google.com>
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Cr-Commit-Position: refs/heads/main@{#1550867}