[ntp-next] Adding deep dive chips
Screenshot: https://screenshot.googleplex.com/AwBa6zLBPzXUfPb Screenshot
for clamped long suggestions:
https://screenshot.googleplex.com/BRswxT5dad3iuWS

This cl uses the css rule `-webkit-box` which is technically deprecated
however since this code will be used for Chrome only, and its
alternative is still in the proposal stage, it should be safe to use
this as its the only choice.

No-Presubmit: True
Reason: False positive on -webkit-box, which is required for -webkit-line-clamp.
Bug: 450023459
Change-Id: I02178e8874538fe45eddccb9ab892448c5a70195
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7205282
Commit-Queue: Omar Flores <omarflores@google.com>
Reviewed-by: Sho Mizoe <shomizoe@google.com>
Reviewed-by: Riley Tatum <rtatum@google.com>
Cr-Commit-Position: refs/heads/main@{#1550873}