[//gpu] Remove gating RasterDecoder ops by client enabling GPU raster
The client is the one that initiates these ops, and the client is also
the one that specifies whether GPU raster is enabled. Hence, there's no
reason to be doing guarding on the service side here.

This is a step toward eliminating service-side dependence on the concept
of GPU raster being enabled.

Bug: 443000825
Change-Id: I7a22bf860add9ca5fdcfcaef26863ea9d6f8480c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7207116
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550880}