[//cc] Add TileBasedLayerImpl::GetIdealContentsScaleKey()
Abstracts the different computations that PictureLayerImpl and
TileDisplayLayerImpl do to produce this key in order to enable followup
code deduplication.

Note: In parallel to the followup deduplication described above, I will
be working to account for this logic [1] in TileDisplayLayerImpl. It is
not clear at this point whether that will result in the maintenance of
the ideal contents scale becoming identical between the two
implementations; if/when that happens, I will eliminate
GetIdealContentsScale() being virtual and instead just dedupe its
implementation directly into TileBasedLayerImpl.

[1] https://source.chromium.org/chromium/chromium/src/+/main:cc/layers/picture_layer_impl.cc;l=1893-1898;drc=db569cd7847d479358bd391a18b2c28393803333

Bug: 450651370
Change-Id: I8721b8baf6cad5c4679439ae366ba563e337c706
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7202604
Reviewed-by: Victor Miura <vmiura@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550884}