Store layer association in the RuleSet (@function, @position-try)
We currently set the CascadeLayer association directly on the rule
as a side effect of RuleSet building. This approach does not work,
as it may lead to rules that are associated with non-existent layers
(see Issue 463698792); we need to store the layer association next to
the rule in the RuleSet instead.

Fixes DCHECK involving unconnected RuleSets and rebuilding of
the cascade layer map (CreateUnconnectedRuleSets_PositionTryRules
and CreateUnconnectedRuleSets_FunctionRules, Issue 428664521).

Unlike other CLs in this chain (CL:7204862, CL:7204203, CL:7206206,
CL:7206226, CL:7205948, CL:7204324, and CL:7207301), this CL fixes
two different at-rules in the same CL (@function and @position-try).
This is because they both use the same AddNameDefiningRules utility,
and temporarily adding two versions of that utility would just add
unnecessary churn.

This is expected to slightly increase memory usage: although a rule
is typically only stored in a single RuleSet, it usually also exists
in at least one other container on ScopedStyleResolver. The assumption
is that we have so few of these rules that the memory impact
is acceptable.

Fixed: 463698792
Issue: 428664521
Change-Id: I2d46718b4bab431b0d8593f83b71c5b6106311a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7207358
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1551020}