[SyncTests] Parameterize TwoClientExtensionsSyncTest
This CL parameterizes TwoClientExtensionsSyncTest to run with different
sync setup modes (kSyncTransportOnly and kSyncTheFeature).

This is part of the broader effort to modernize sync integration tests
to ensure higher coverage.

Additionally, this CL enables `StartWithDifferentExtensions` and
`UninstallPreinstalledExtensions` for `kSyncTransportOnly` mode. Extensions installed locally before sign-in are not automatically uploaded to the account in transport mode. The tests now explicitly trigger the upload of local extensions to the account using `extensions::sync_util::UploadExtensionToAccount` when running in transport mode, mimicking the batch upload behavior.

Bug: 353425612
Change-Id: I3ff5b4b357a3efdf3859f28564d2724b416fd9e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7208318
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Michael Tatarski <mtatarski@google.com>
Cr-Commit-Position: refs/heads/main@{#1551051}