[FontDataService] Enable FontDataService/Manager on Linux
A new feature is added to enable FontDataService/Manager on Linux. This
comes as we have ramped up the Windows feature to Stable 10%.

Some unit tests were not enabled on Linux due to not supported
operations on the default SkFontManager, and differences in how
GetResourceName() is handled causing FontDataService to always use the
memory region approach. These will be addressed later.

The FontLoader on RendererBlinkPlatformImpl should also be removed later
and replaced by FontDataManager if the feature is on.

Bug: 458737348
Change-Id: I4e523125231ba726426bd42cbb5763dff6b0779d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7079251
Reviewed-by: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Andres Ricardo Perez <andresrperez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1551080}