[iOS]Replace DCHECK by CHECK in settings
Some bugs related to the settings are harder to debug because all
relevant CHECK that could show the cause of the issue are actuall
DCHECK. Replacing them by CHECK, that will be fatal in many milestones,
will help find out those corner cases.

As an example of such bug, the remaining case of crbug.com/438409485 is
caused by the fact that the profil is nil during viewDidLoad. It is
impossible to know whether it’s nil because the view controller was
dismissed or whether the profile has never been initialized because the
browser itself was nil.

Bug: 438409485
Change-Id: Ie816c5faa3af6c5de450a79e7947df13f2b7f6a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7207339
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1551085}