Ensure TransferableResource is released in tests
This CL adds manual release of TransferableResource is several tests.
The reason for this change is that in the upcoming replacement of
Mailbox with ClientSharedImage in TransferableResource, we need to
fully destroy the resource to ensure its strong ClientSharedImage ref
is released, before the ClientSharedImage itself can be destructed.

Specifically, this CL would fix 4 tests that would have otherwise
failed once the ClientSharedImage refactoring is introduced:

DrawingBufferImageChromiumTest.VerifyResizingReallocatesImages
DrawingBufferTest.VerifySharedImagesReleasedAfterReleaseCallback
WebGPUSwapBufferProviderTest.VerifyDestructionCompleteAfterAllResourceReleased
WebGPUSwapBufferProviderTest.VerifyInsertAndWaitSyncTokenCorrectly

Bug: 40286368
Change-Id: I4bb7038002151fdb97a0d7fbb78f9056f6df9ae5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7204949
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1551140}