[Chrome Next] Fix superfluous separator in certain Mac fullscreen modes
When in immersive mode with "always show toolbar" on Mac, whether a
separator is shown should be dependent on whether there is a shadow
box around the contents.

This CL also fixes a bug where infobars inside the shadowbox would slide
down with immersive mode reveal. (As a follow-up, we can instead apply
this slide to the entire shadow box and its contents if needed.)

Bug: 451658294
Change-Id: I07c0ab907385d30ce1761ea256071ae73f2b6c09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7207987
Reviewed-by: Vince Lugli <lugli@google.com>
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1551151}