legion: Add UMA metrics for Legion Client request latency
This change introduces a new UMA histogram,
"Legion.Client.RequestLatency.{Status}", to track the time taken for
requests sent from the Legion Client. The latency is measured from when
the request is sent until a response is received, the request times out,
or an error occurs. The "{Status}" suffix indicates whether the request
resulted in Success, Timeout, or Error.

Bug: b:457652337
Change-Id: Id84ccdf02c83a959cda1bf0c0799cdaedaa82f04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7203025
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Nikhil Jakhar <nikhiljakhar@google.com>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1551153}