Add is_software to ExportedSharedImage
This CL adds an additional is_software field to ExportedSharedImage to
support the replacement of Mailbox in TransferableResource with
ClientSharedImage. This change is intended to fix failures in the
GLAndSoftwareCompositing/CompositingRenderWidgetHostViewBrowserTest*
and HeadlessCompositorBrowserTest* that would have otherwise occurred
when the ClientSharedImage is introduced in TransferableResource.

Bug: 40286368
Change-Id: Ifc105af902ad317aa2e67e707b6b29ebaeb1aa49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7204658
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1551205}